# Optimize json.Compact and json.Indent