## Benchmark

around **2x** faster than encoding/json

```
$ go test -bench Indent
goos: darwin
goarch: amd64
pkg: benchmark
Benchmark_Indent_EncodingJson-16              38          30272302 ns/op        17069782 B/op         19 allocs/op
Benchmark_Indent_GoJson-16                    80          14656825 ns/op         5169307 B/op          2 allocs/op
PASS
ok      benchmark       3.615s

$ go test -bench Compact
goos: darwin
goarch: amd64
pkg: benchmark
Benchmark_Compact_EncodingJson-16            100          10036194 ns/op         1941742 B/op          1 allocs/op
Benchmark_Compact_GoJson-16                  190           6241213 ns/op         1941677 B/op          1 allocs/op
PASS
ok      benchmark       3.088s

```